レポート機能の処理変更＋α

+++ model/deleteIncByTrans.php
@@ -8,6 +8,7 @@
  * @access public
  * @package model
  * @name deleteIncByTrans
+ * @var object $model モデルクラス共通処理オブジェクト
  * @var int $userID ユーザID
  * @var int $id 収入情報ID
  * @var array $result クエリ実行結果
@@ -22,7 +23,6 @@ class deleteIncByTrans
     
     /**
      * コンストラクタ
-     * 何もしない
      *
      * @access public
      */
@@ -58,16 +58,55 @@ class deleteIncByTrans
             // DB接続情報取得
             include 'tools/databaseConnect.php';
             
-            // 収入情報の削除
-            $query = "DELETE FROM incomeTable WHERE incomeID = '$id' AND userID = '$userID'";
-            $queryResult = mysqli_query($link, $query);
-            $this->result = mysqli_fetch_assoc($queryResult);
-            
+            // DB接続に失敗した場合
+            if ($link == false) {
+                $DBConnect = "failed";
+                $this->model -> setDBConnectResult($DBConnect);
+                $this->result = null;
+                
+            } else {
+                $DBConnect = "success";
+                $this->model -> setDBConnectResult($DBConnect);
+                
+                // 事前確認
+                $query = "
+                    SELECT COUNT(*)
+                    FROM incomeTable
+                    WHERE userID = '$userID'
+                    ";
+                $queryResult = mysqli_query($link, $query);
+                $result = mysqli_fetch_assoc($queryResult);
+                $countBefore = (int) $result['COUNT(*)'];
+                
+                // 収入情報の削除
+                $query = "DELETE FROM incomeTable WHERE incomeID = '$id' AND userID = '$userID'";
+                $queryResult = mysqli_query($link, $query);
+                $this->result = mysqli_fetch_assoc($queryResult);
+                
+                // 事後確認
+                $query = "
+                    SELECT COUNT(*)
+                    FROM incomeTable
+                    WHERE userID = '$userID'
+                    ";
+                $queryResult = mysqli_query($link, $query);
+                $result = mysqli_fetch_assoc($queryResult);
+                $countAfter = (int) $result['COUNT(*)'];
+                
+                $countDiff = $countAfter - $countBefore;
+                
+                if ($countDiff == -1) {
+                    $this->result = true;
+                    
+                } else {
+                    $this->result = false;
+                    
+                }
+            }
             // DB切断
             mysqli_close($link);
         
         }
-        
         return $this->result;
         
     }
